requirements.txt
Python==3.10.5
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0
drf-spectacular==0.24.2